+++ prelude/src/time.rs
@@ -1,7 +1,7 @@
-pub const HOUR: i64 = 60 * 60;
-pub const DAY: i64 = HOUR * 24;
-pub const WEEK: i64 = DAY * 7;
-pub const MONTH: i64 = WEEK * 4;
+pub const HOUR: u64 = 60 * 60;
+pub const DAY: u64 = HOUR * 24;
+pub const WEEK: u64 = DAY * 7;
+pub const MONTH: u64 = WEEK * 4;
 
 pub fn current_timestamp() -> u64 {
     std::time::SystemTime::now()
@@ -12,17 +12,17 @@ pub fn current_timestamp() -> u64 {
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq, serde::Deserialize, serde::Serialize)]
 pub struct AbsoluteTimeRange {
-    pub start: i64,
+    pub start: u64,
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub end: Option<i64>,
+    pub end: Option<u64>,
 }
 
 impl AbsoluteTimeRange {
-    pub fn since(start: i64) -> Self {
+    pub fn since(start: u64) -> Self {
         Self { start, end: None }
     }
 
-    pub fn until(mut self, end: i64) -> Self {
+    pub fn until(mut self, end: u64) -> Self {
         self.end = Some(end);
         self
     }
@@ -38,7 +38,7 @@ pub enum RelativeTimeRange {
 }
 
 impl RelativeTimeRange {
-    pub const fn duration(&self) -> i64 {
+    pub const fn duration(&self) -> u64 {
         match self {
             Self::LastHour => HOUR,
             Self::LastDay => DAY,
@@ -48,7 +48,7 @@ impl RelativeTimeRange {
     }
 
     pub fn into_absolute(self) -> AbsoluteTimeRange {
-        let now = current_timestamp() as i64;
+        let now = current_timestamp();
         AbsoluteTimeRange {
             start: now - self.duration(),
             end: None,
